requirements.txt
tk
pyperclip
Pillow

